added dtos
