--- src/main/java/org/braekpo1nt/mctmanager/games/game/config/EnchantmentDTO.java
@@ -1,10 +1,0 @@
-package org.braekpo1nt.mctmanager.games.game.config;
-
-import org.bukkit.enchantments.Enchantment;
-import org.bukkit.enchantments.EnchantmentWrapper;
-
-public record EnchantmentDTO(String name, int level) {
-    public Enchantment toEnchantment() {
-        return new EnchantmentWrapper(name);
-    } 
-}
